Creating the different fractions and calling the constructors. Fractions updated to correct the use of double in the division.

+++ prepare/Learning03/Fraction.cs
@@ -45,6 +45,6 @@ class Fraction
 
     public double GetDecimalValue()
     {
-        return _top/_bottom;
+        return (double)_top/(double)_bottom;
     }
 }
